timeframe-eval/requirements.txt
mmif
pyannote.metrics
glob
argparse
moviepy
